requirements.txt
pyfuse3
dnspython